LayoutTests/fast/forms/validation-message-on-textarea.html

<!DOCTYPE html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p>Check if the position of a validation message for a textarea is good.</p>

<div style="padding: 20px"> <!-- The containing block of the target element is not the body. -->
<form action="validation-message-on-textarea.html">
<textarea id=host required style="margin: 31px;"></textarea>
<input type=submit id=submit>
</form>
</div>

<div id="console"></div>
<script>
function check() {
    hostRect = getAbsoluteRect(host);
    bubbleRect = getAbsoluteRect(getValidationMessageBubbleNode(host));
    debug('The top of the bubble should be on the bottom of the host node:');
    shouldBe('bubbleRect.top', 'hostRect.bottom');
    debug('');

    finishJSTest();
}

var host = $('host');
var hostRect;
var bubbleRect;
$('submit').click();
if (window.testRunner) {
    setTimeout(check, 0);
} else {
    debug('The test requires testRunner.');
}

var jsTestIsAsync = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
