LayoutTests/fast/forms/validation-message-on-textarea-expected.txt

Check if the position of a validation message for a textarea is good.


The top of the bubble should be on the bottom of the host node:
PASS bubbleRect.top is hostRect.bottom

PASS successfullyParsed is true

TEST COMPLETE
